+++ packages/eslint-plugin/tests/rules/no-unused-vars/no-unused-vars-eslint.test.ts
@@ -4,10 +4,7 @@
 
 'use strict';
 
-import {
-  AST_NODE_TYPES,
-  TSESLint,
-} from '@typescript-eslint/experimental-utils';
+import { AST_NODE_TYPES, TSESLint } from '@typescript-eslint/utils';
 import rule, { MessageIds } from '../../../src/rules/no-unused-vars';
 import { RuleTester } from '../../RuleTester';